CrudAuth filter as a mandatory filter in query
This fixes the issue explained in #323 and #302. 

For use case of explained issue it works fine. Only concern is that this line:

const defaultSearch = this.getDefaultSearchCondition(options, parsed);

not only gets parsed.authFilter, but also options.query.filter and parsed.paramsFilter. Because of that I don't know if the fix can affect this two parameters in an unexpected behavior. Review necessary by a project author for other use cases.