feat: Customizeable CrudRoutesFactory
CrudRoutesFactory has a number of important elements related to Swagger generation that is generally hidden from the called.  By providing a subclass of CrudRoutesFactory to @Crud options, one is able to expose hooks to allow customization of Swagger documentation.

The author is primarily concerned with being able to customize the opererationId to reduce the length of identifiers used by common Code-generation tools.